+++ src/main/cz/fel/ds/database/model/MealToFood.java
@@ -1,7 +1,6 @@
 package cz.fel.ds.database.model;
 
-import javafx.beans.property.SimpleFloatProperty;
-
+import javafx.beans.property.SimpleDoubleProperty;
 import javax.persistence.*;
 import java.io.Serializable;
 
@@ -11,34 +10,37 @@ import java.io.Serializable;
 @Entity
 @Table(name = "MealToFood")
 public class MealToFood implements Serializable {
-    private SimpleFloatProperty quantity = new SimpleFloatProperty();
+    private SimpleDoubleProperty quantity = new SimpleDoubleProperty();
     private Meal meal;
     private Food food;
 
     @Column(name = "quantity", length = 100)
-    public float getQuantity() {
+    public double getQuantity() {
         return quantity.get();
     }
-    public void setQuantity(float quantity) {
+
+    public void setQuantity(double quantity) {
         this.quantity.set(quantity);
     }
 
     @Id
     @OneToOne(cascade = CascadeType.ALL)
-    @JoinColumn(name="food_id")
+    @JoinColumn(name = "food_id")
     public Food getFood() {
         return food;
     }
+
     public void setFood(Food food) {
         this.food = food;
     }
 
     @Id
     @OneToOne(cascade = CascadeType.ALL)
-    @JoinColumn(name="meal_id")
+    @JoinColumn(name = "meal_id")
     public Meal getMeal() {
         return meal;
     }
+
     public void setMeal(Meal meal) {
         this.meal = meal;
     }
